change kafka topic config int value from int64
API retrieves values like 9223372036854776000, which is out of range for int

+++ kafka_topic_acc_test.go
@@ -74,7 +74,7 @@ var _ = Describe("Kafka Topic", func() {
 		var (
 			errC            error
 			topicName       string
-			segmentJitterMs int
+			segmentJitterMs int64
 		)
 		segmentJitterMs = 10
 		topicName = "test1"
